refactor: add folder pages, change structure
