+++ frontend/src/index.js
@@ -1,32 +1,19 @@
 import React from "react";
-import ReactDOM from "react-dom/client";
-import {
-  createBrowserRouter,
-  RouterProvider,
-  Routes,
-  Route,
-  Router,
-} from "react-router-dom";
+import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import reportWebVitals from "./reportWebVitals";
-import LoginRoute from "./routes/LoginRoute";
-import CalendarRoute from "./routes/CalendarRoute";
+import ReactDOM from "react-dom/client";
 
-import Header from "./components/Header/Header";
+import LoginRoute from "./pages/LoginRoute";
+import CalendarRoute from "./pages/CalendarRoute";
+
 import "./index.css";
-import ModeusLoginForm from "./components/Login/ModeusLoginForm";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 const router = createBrowserRouter([
-  {
-    path: "/",
-    element: (
-      <div className="wrapper">
-        <Header />
-        <div className="main-calendar">
-          <CalendarRoute />
-        </div>
-      </div>
-    ),
+    //Todo: если не авторизаван тогда перекидывать на авторизацию
+    {
+     path: "/",
+     element: <CalendarRoute />,
   },
   {
     path: "/login",
@@ -41,19 +28,7 @@ const router = createBrowserRouter([
 root.render(
   <React.StrictMode>
     <RouterProvider router={router} />
-  </React.StrictMode>,
+  </React.StrictMode>
 );
-
-const App = () => {
-  return (
-    <dev>
-      <Router>
-        <Routes>
-          <Route path="/" element={<LoginRoute />} />
-        </Routes>
-      </Router>
-    </dev>
-  );
-};
 
 reportWebVitals();
