+++ frontend/src/pages/CalendarRoute.jsx
@@ -1,9 +1,10 @@
-import Calendar from "../components/Calendar/Calendar"
 import DataPicker from "../components/Calendar/DataPicker"
+import Header from "../components/Header/Header";
 
 const CalendarRoute = () => {
     return (
         <div>
+            <Header />
             <DataPicker/>
         </div>
     )
