+++ frontend/src/services/api/login.js
@@ -19,6 +19,7 @@ export async function searchModeus(fullName) {
         const params = new URLSearchParams();
         params.append('full_name', fullName);
         return await axios.get(`${BACKEND_URL}/api/modeus/search_blank/`, params);
+
     } catch (e) {
         return e.response;
     }
